backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
sqlalchemy==2.0.28
alembic==1.13.1
pydantic==2.6.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
python-dotenv==1.0.1
httpx==0.27.0
# SQLite is built into Python — no extra driver needed
# psycopg2-binary  ← add this back if you switch to PostgreSQL in production
